Booyeah! I had to correct for the new TSS.ESP0 position (i.e. needed it to go back to the end of the stack) - without correcting, it'd keep going forward (!) because we keep setting it to the start of the new position, not the end.
I mean; we first set TSS.ESP0 to the `top' (end) of the ukts. When we return from that task, it's written down the stack a bit, and we record that position so we can get to its data. But then we were also setting TSS.ESP0 to be
that new position - which means next time we come around, the TSS would start writing there, below our old stack. That's useless, and means after a few iterations, we'd write our way write out of the ukts and into who-knows
land. We want to keep rewriting the same position in the ukts. We do that now. :)